+++ terraform/hydrocron-iam.tf
@@ -70,7 +70,8 @@ data "aws_iam_policy_document" "dynamo-write-policy" {
 
     resources = [
       aws_dynamodb_table.hydrocron-swot-node-table.arn,
-      aws_dynamodb_table.hydrocron-swot-reach-table.arn
+      aws_dynamodb_table.hydrocron-swot-reach-table.arn,
+      aws_dynamodb_table.hydrocron-swot-prior-lake-table.arn
     ]
   }